Update Minneapolis.m
I noticed that the variable dec is initially set to the averages in table 1.6.1, but is later changed to different values in line 124. Not sure if this was intentional, but regardless I had to redefine the average declinations for the year to dec2. I noted this in the code as well.